rename some variables
